Move kafkaClusterId to WorkerConfig
